+++ package.json
@@ -11,6 +11,7 @@
     "@types/node": "^16.18.11",
     "@ya.praktikum/react-developer-burger-ui-components": "^1.14.0",
     "immutability-helper": "^3.1.1",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
@@ -21,6 +22,7 @@
     "redux": "^4.2.1",
     "redux-thunk": "^2.4.2",
     "typescript": "^4.9.4",
+    "uuid": "^9.0.0",
     "uuidv4": "^6.2.13",
     "web-vitals": "^2.1.4"
   },
