+++ src/components/burger-constructor/burger-constructor.jsx
@@ -1,305 +1,8 @@
-import React, {useMemo, useRef} from "react";
-import PropTypes from "prop-types";
-import {
-    ConstructorElement,
-    DragIcon,
-    Button,
-    CurrencyIcon,
-} from "@ya.praktikum/react-developer-burger-ui-components";
-import {OrderDetails} from "../order-details/order-details"
-import {Modal} from "../modal/modal";
+import React from "react";
 import burgerConstructorStyle from "./burger-constructor.module.css";
-import {ingredientType} from "../../utils/prop-types";
-import {useDispatch, useSelector} from 'react-redux';
-import {
-    ADD_COMPONENT,
-    CHANGE_BUN,
-    REMOVE_COMPONENT,
-    SORT_COMPONENT,
-    CLEAR_COMPONENTS,
-} from "../../services/actions/constructor-burger"
-import {ORDER_CLEAR, getOrder} from "../../services/actions/order";
-import {useDrag, useDrop} from "react-dnd";
-import update from "immutability-helper";
-import {v4 as uuidv4} from "uuid";
-
-const ItemConstructor = ({ingredient, index, itemKey}) => {
-    const dispatch = useDispatch();
-    const ref = useRef(null);
-    const componentsData = useSelector((store) => store.burgerConstruct);
-
-
-    const moveCard = (dragIndex, hoverIndex) => {
-        const components = componentsData.components;
-        const newComp = components.slice();
-
-        const spliced = update(newComp, {
-            $splice: [
-                [dragIndex, 1],
-                [hoverIndex, 0, newComp[dragIndex]],
-            ],
-        });
-        dispatch({
-            type: SORT_COMPONENT,
-            components: spliced,
-        });
-    }
-
-    const [{opacity}, dragRef] = useDrag({
-        type: "component",
-        item: () => {
-            return {ingredient, index};
-        },
-        collect: (monitor) => ({
-            opacity: monitor.isDragging() ? 0.5 : 1,
-        }),
-    });
-
-    const [{handlerId}, dropRef] = useDrop({
-        accept: "component",
-        collect(monitor) {
-            return {
-                handlerId: monitor.getHandlerId(),
-            };
-        },
-        hover(item, monitor) {
-            if (!ref.current) {
-                return;
-            }
-            // console.log(index);
-            const dragIndex = item.index;
-            const hoverIndex = index;
-            // Don't replace items with themselves
-            if (dragIndex === hoverIndex) {
-                return;
-            }
-            // Determine rectangle on screen
-            const hoverBoundingRect = ref.current?.getBoundingClientRect();
-            // Get vertical middle
-            const hoverMiddleY =
-                (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-            // Determine mouse position
-            const clientOffset = monitor.getClientOffset();
-            // Get pixels to the top
-            const hoverClientY = clientOffset.y - hoverBoundingRect.top;
-            // Dragging downwards
-            if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-                return;
-            }
-            // Dragging upwards
-            if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-                return;
-            }
-            moveCard(dragIndex, hoverIndex);
-            item.index = hoverIndex;
-        },
-        drop(item) {
-            const dragIndex = item.index;
-            const hoverIndex = index;
-            if (dragIndex === hoverIndex) {
-                return;
-            }
-            moveCard(dragIndex, hoverIndex);
-        },
-    });
-
-    dragRef(dropRef(ref));
-
-    const deleteCard = () => {
-        dispatch({type: REMOVE_COMPONENT, key: itemKey});
-    };
-
-
-    return (
-        <li className={"pl-8 " + burgerConstructorStyle.card} ref={ref} data-handler-id={handlerId}>
-            <div className={burgerConstructorStyle.drag}>
-                <DragIcon type="primary"/>
-            </div>
-            <div className={burgerConstructorStyle.element} style={{opacity}}>
-                <ConstructorElement
-                    type={ingredient.type}
-                    text={ingredient.name}
-                    price={ingredient.price}
-                    thumbnail={ingredient.image}
-                    handleClose={() => deleteCard()}
-                />
-            </div>
-        </li>
-    );
-};
-
-const Plug = ({type}) => {
-    return (
-        <li className={"pl-8 " + burgerConstructorStyle.card}>
-            <div className={"text text_type_main-small " + burgerConstructorStyle.plug}>
-                {/*<div className={"text text_type_main-small " + burgerConstructorStyle.plug__text}>*/}
-                {type !== 'bun' ?
-                    'Выберите начинку'
-                    :
-                    'Выберите булки'
-                }
-                {/*</div>*/}
-            </div>
-        </li>
-    );
-};
-
-const ItemConstructorLocked = ({ingredient, position}) => {
-    return (
-        <li className="pl-8">
-            <ConstructorElement
-                isLocked={true}
-                type={ingredient.type}
-                text={ingredient.name + position}
-                price={ingredient.price / 2}
-                thumbnail={ingredient.image}
-
-            />
-        </li>
-    );
-};
-
-
-const ConstructorBox = ({ingredients}) => {
-    const dispatch = useDispatch();
-    const [, dropTarget] = useDrop({
-        accept: "ingredient",
-        // collect: (monitor) => ({
-        //     isHover: monitor.isOver(),
-        // }),
-        drop({card}) {
-            if (card.type === "bun") {
-                dispatch({type: CHANGE_BUN, id: card._id});
-            } else {
-                let key = uuidv4();
-                dispatch({type: ADD_COMPONENT, id: card._id, key: key});
-            }
-        },
-    });
-
-    const ingredientsData = useSelector(state => state.burger.ingredients)
-
-
-    const components = useSelector((store) => store.burgerConstruct);
-
-
-    return (
-        <ul
-            className={burgerConstructorStyle.box}
-            ref={dropTarget}
-        >
-            {components.bun ? ingredientsData.map(
-                (item) =>
-                    item._id === components.bun && (
-                        <ItemConstructorLocked
-                            key={item._id}
-                            ingredient={item}
-                            position={" (верх)"}
-                        />
-                    )
-            ) : (<Plug type='bun'/>
-            )}
-            <li>
-                <ul className={burgerConstructorStyle.box_active}>
-                    {ingredients.length === 1 ? (
-                        <Plug type='main'/>
-                    ) : (
-                        components.components.map(({id, key}, index) => {
-                            const ingredient = ingredientsData.find(
-                                (el) => el._id === id && el.type !== "bun"
-                            );
-                            return (
-                                ingredient && (
-                                    <ItemConstructor
-                                        key={key}
-                                        ingredient={ingredient}
-                                        index={index}
-                                        itemKey={key}
-                                    />
-                                )
-                            );
-                        })
-                    )}
-                </ul>
-            </li>
-            {components.bun ? ingredientsData.map(
-                (item) =>
-                    item._id === components.bun && (
-                        <ItemConstructorLocked
-                            key={item._id}
-                            ingredient={item}
-                            position={" (низ)"}
-                        />
-                    )
-            ) : (<Plug type='bun'/>
-            )}
-
-        </ul>
-
-
-    );
-};
-
-const ConstructorBoxPrice = ({ingredients}) => {
-    const ingredientsData = useSelector((state) => state.burger.ingredients);
-
-
-    const price = useMemo(() => {
-        let total = 0;
-        ingredients.forEach((item) => {
-            const ingredient = ingredientsData.find((el) => el._id === item);
-            if (ingredient) {
-                total += ingredient.price;
-            }
-        });
-        return total;
-    }, [ingredients, ingredientsData]);
-
-    return (
-        <div className={"mr-10 " + burgerConstructorStyle.price}>
-            <p className="text text_type_digits-medium">{price}</p>
-            <CurrencyIcon type="primary"/>
-        </div>
-    );
-};
-
-const ConstructorButtonBox = ({ingredients}) => {
-    const dispatch = useDispatch()
-    const {order} = useSelector(store => store.order)
-
-
-    const handleClose = () => {
-        dispatch({
-            type: ORDER_CLEAR
-        });
-        dispatch({
-            type: CLEAR_COMPONENTS,
-        });
-    };
-
-
-    const modal =
-        <Modal onClose={handleClose}>
-            <OrderDetails/>
-        </Modal>;
-
-    return (
-        <div className={"mr-4 mt-10 " + burgerConstructorStyle.button_order}>
-            <ConstructorBoxPrice ingredients={ingredients}/>
-            <Button
-                htmlType="button"
-                type="primary"
-                size="large"
-                onClick={async () => {
-                    dispatch(getOrder(ingredients));
-                }}
-            >
-                Оформить заказ
-            </Button>
-            {order && modal}
-        </div>
-    );
-};
+import { useSelector} from 'react-redux';
+import {ConstructorButtonPriceBox} from "./constructor-button-price-box";
+import {ConstructorBox} from "./constructor-box";
 
 
 export const BurgerConstructor = () => {
@@ -327,32 +30,7 @@ export const BurgerConstructor = () => {
     return (
         <section className={"pt-25 " + burgerConstructorStyle.constructor}>
             <ConstructorBox ingredients={ingredients}/>
-            <ConstructorButtonBox ingredients={ingredients}/>
+            <ConstructorButtonPriceBox ingredients={ingredients}/>
         </section>
     );
 }
-
-ItemConstructor.propTypes = {
-    ingredient: ingredientType.isRequired,
-    index: PropTypes.number.isRequired,
-    itemKey: PropTypes.string.isRequired
-};
-
-
-ItemConstructorLocked.propTypes = {
-    ingredient: ingredientType.isRequired,
-    position: PropTypes.string.isRequired,
-};
-
-ConstructorBoxPrice.propTypes = {
-    ingredients: PropTypes.arrayOf(PropTypes.string).isRequired,
-}
-
-
-ConstructorBox.propTypes = {
-    ingredients: PropTypes.arrayOf(PropTypes.string).isRequired,
-};
-
-ConstructorButtonBox.propTypes = {
-    ingredients: PropTypes.arrayOf(PropTypes.string).isRequired
-};
