+++ src/components/burger-constructor/burger-constructor.module.css
@@ -11,25 +11,6 @@
   gap: 16px;
 }
 
-.box_scroll {
-  composes: box;
-  max-height: 470px;
-  overflow-y: auto;
-  overflow-x: hidden;
-  scrollbar-width: thin;
-  scrollbar-color: #8585AD #2F2F37;
-}
-
-.box_scroll::-webkit-scrollbar {
-  background-color: #2F2F37;
-  width: 8px;
-
-}
-
-.box_scroll::-webkit-scrollbar-thumb {
-  background-color: #8585AD;
-}
-
 .card {
   position: relative;
 }
@@ -76,14 +57,26 @@
   background-color: #8585AD;
 }
 
-.plug {
+
+.plug{
   width: 536px;
   height: 80px;
   display: flex;
   align-items: center;
   justify-content: center;
   background: #1C1C21;
+}
+
+.main {
   border-radius: 40px;
 }
 
+.top {
+  border-radius: 88px 88px 40px 40px;
 
+}
+
+.bottom {
+  border-radius: 40px 40px 88px 88px;
+
+}
