+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -80,3 +80,7 @@
 .box::-webkit-scrollbar-thumb {
   background-color: #8585AD;
 }
+
+.tab{
+  display: flex;
+}
