+++ src/services/actions/order.jsx
@@ -1,4 +1,5 @@
 import {Burger_API} from "../../utils/burger-api";
+import {checkResponse} from "../../utils/check-response";
 
 export const GET_ORDER_REQUEST = "GET_ORDER_REQUEST";
 export const GET_ORDER_SUCCESS = "GET_ORDER_SUCCESS";
@@ -17,12 +18,7 @@ export function getOrder(ingredients) {
                 ingredients: ingredients,
             }),
         })
-            .then(function (res) {
-                if (res.ok) {
-                    return res.json();
-                }
-                return Promise.reject(`Ошибка: ${res.statusText}`);
-            })
+            .then(checkResponse)
             .then((res) => {
                 if (res && res.success) {
                     dispatch({
