+++ src/utils/burger-api.js
@@ -1,5 +1,3 @@
-import React from "react";
-
 export const Burger_API = "https://norma.nomoreparties.space/api";
