+++ src/utils/check-response.js
@@ -1,0 +1,6 @@
+export function checkResponse(res) {
+        if (res.ok) {
+            return res.json();
+        }
+        return Promise.reject(`Ошибка: ${res.statusText}`);
+}
